Added pomodoro timer, styled more of the home page
In this interval of working on this project, I am in the midst of building the pomodoro timer and styling the home page.

Compromises from my original design have been made - all documented in my README file.

The timer requires de-bugging and functionality is not complete. The other elements are in very early stages.