Developed nav bar, pomodoro timer and kanban board. Conducted user testing and iterated on feedback.
In this next phase of the development process, I conducted A/B user testing to compare different designs of components such as the Pomodoro Timer and Kanban Board. I analysed the feedback and quotes here to align back to user goals as defined in A2. Evidence of this can be seen in the README.md file.

I also received multiple pieces of feedback from tutors in my development process as I sharpened my understanding of HTML, CSS and Javascript.